Made ProductService async; Added ProductsController;

+++ AdventureWorks.Services/ProductsService.cs
@@ -4,53 +4,56 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Threading.Tasks;
 
 namespace AdventureWorks.Services
 {
     public class ProductsService : IProductsService
     {
-        public Product Create(Product product)
+        public async Task<Product> Create(Product product)
         {
             using (var db = new DataContext())
             {
-                db.Product.Add(product);
-                db.SaveChanges();
+                await db.Product.AddAsync(product);
+                await db.SaveChangesAsync();
 
                 return product;
             }
         }
 
-        public IEnumerable<Product> GetAll()
+        public async Task<IEnumerable<Product>> GetAll()
         {
             using (var db = new DataContext())
             {
-                return db.Product.ToList();
+                return await db.Product.ToListAsync();
             }
         }
 
-        public Product GetById(int id)
+        public async Task<Product> GetById(int id)
         {
             using (var db = new DataContext())
             {
-                return db.Product.Find(id);
+                return await db.Product.FindAsync(id);
             }
         }
 
-        public void Remove(Product product)
+        public async Task Remove(Product item)
         {
             using (var db = new DataContext())
             {
+                var product = await db.Product.FindAsync(item.ProductId);
+                if (product == null) return;
                 db.Product.Remove(product);
-                db.SaveChanges();
+                await db.SaveChangesAsync();
             }
         }
 
-        public Product Update(Product product)
+        public async Task<Product> Update(Product product)
         {
             using (var db = new DataContext())
             {
                 db.Entry(product).State = EntityState.Modified;
-                db.SaveChanges();
+                await db.SaveChangesAsync();
 
                 return product;
             }
